Backend dashboard metrics & stackover issue resolved

+++ src/main/java/com/project/skillswap/logic/entity/LearningSession/LearningSessionRepository.java
@@ -2,5 +2,5 @@ package com.project.skillswap.logic.entity.LearningSession;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LearningSessionRepository extends JpaRepository<LearningSession, Integer> {
+public interface LearningSessionRepository extends JpaRepository<LearningSession, Long> {
 }
